refactor executor scenario package
