+++ examples/http/bench_http.go
@@ -8,7 +8,7 @@ import (
 	"time"
 
 	httpClient "github.com/gobench-io/gobench/clients/http"
-	"github.com/gobench-io/gobench/scenario"
+	"github.com/gobench-io/gobench/executor/scenario"
 )
 
 func Export() scenario.Vus {
