+++ examples/mqtt/1_to_1/1_to_1.go
@@ -16,7 +16,7 @@ import (
 
 	"github.com/gobench-io/gobench/clients/mqtt"
 	"github.com/gobench-io/gobench/dis"
-	"github.com/gobench-io/gobench/scenario"
+	"github.com/gobench-io/gobench/executor/scenario"
 )
 
 func Export() scenario.Vus {
