+++ examples/mqtt/fan_out/fan_out.go
@@ -16,7 +16,7 @@ import (
 
 	"github.com/gobench-io/gobench/clients/mqtt"
 	"github.com/gobench-io/gobench/dis"
-	"github.com/gobench-io/gobench/scenario"
+	"github.com/gobench-io/gobench/executor/scenario"
 )
 
 func Export() scenario.Vus {
