+++ examples/mqtt/query/query.go
@@ -22,7 +22,7 @@ import (
 
 	"github.com/gobench-io/gobench/clients/mqtt"
 	"github.com/gobench-io/gobench/dis"
-	"github.com/gobench-io/gobench/scenario"
+	"github.com/gobench-io/gobench/executor/scenario"
 )
 
 const (
