+++ examples/nats/bench_nats.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/gobench-io/gobench/clients/nats"
 	"github.com/gobench-io/gobench/dis"
-	"github.com/gobench-io/gobench/scenario"
+	"github.com/gobench-io/gobench/executor/scenario"
 )
 
 func Export() scenario.Vus {
